+++ src/SmartTalk.Core/Services/AiSpeechAssistant/AiSpeechAssistantService.cs
@@ -8,8 +8,8 @@ using System.Net.WebSockets;
 using Microsoft.AspNetCore.Http;
 using Smarties.Messages.Extensions;
 using SmartTalk.Core.Settings.OpenAi;
+using SmartTalk.Messages.Commands.AiSpeechAssistant;
 using SmartTalk.Messages.Enums.OpenAi;
-using SmartTalk.Messages.Commands.PhoneCall;
 using SmartTalk.Messages.Dto.AiSpeechAssistant;
 using SmartTalk.Messages.Events.AiSpeechAssistant;
 using Twilio.TwiML.Voice;
